refactor: use lowercase naming
Rename files to be lowercase and use hyphens as separator.
