+++ administrative-sdk/controllers/pronunciation-challenge-controller.js
@@ -1,4 +1,4 @@
-const PronunciationChallenge = require('../models/pronunciationChallenge');
+const PronunciationChallenge = require('../models/pronunciation-challenge');
 
 /**
  * Controller class for the PronunciationChallenge model.
