+++ administrative-sdk/models/speech-challenge.js
