+++ administrative-sdk/models/speech-recording.js
