+++ index.js
@@ -1,25 +1,25 @@
-const BasicAuth = require('./administrative-sdk/models/basicAuth');
-const BasicAuthController = require('./administrative-sdk/controllers/basicAuthController');
-const ChoiceChallenge = require('./administrative-sdk/models/choiceChallenge');
-const ChoiceChallengeController = require('./administrative-sdk/controllers/choiceChallengeController');
-const ChoiceRecognition = require('./administrative-sdk/models/choiceRecognition');
-const ChoiceRecognitionController = require('./administrative-sdk/controllers/choiceRecognitionController');
-const Connection = require('./administrative-sdk/controllers/connectionController');
+const BasicAuth = require('./administrative-sdk/models/basic-auth');
+const BasicAuthController = require('./administrative-sdk/controllers/basic-auth-controller');
+const ChoiceChallenge = require('./administrative-sdk/models/choice-challenge');
+const ChoiceChallengeController = require('./administrative-sdk/controllers/choice-challenge-controller');
+const ChoiceRecognition = require('./administrative-sdk/models/choice-recognition');
+const ChoiceRecognitionController = require('./administrative-sdk/controllers/choice-recognition-controller');
+const Connection = require('./administrative-sdk/controllers/connection-controller');
 const Organisation = require('./administrative-sdk/models/organisation');
-const OrganisationController = require('./administrative-sdk/controllers/organisationController');
+const OrganisationController = require('./administrative-sdk/controllers/organisation-controller');
 const Phoneme = require('./administrative-sdk/models/phoneme');
-const PronunciationAnalysis = require('./administrative-sdk/models/pronunciationAnalysis');
-const PronunciationAnalysisController = require('./administrative-sdk/controllers/pronunciationAnalysisController');
-const PronunciationChallenge = require('./administrative-sdk/models/pronunciationChallenge');
-const PronunciationChallengeController = require('./administrative-sdk/controllers/pronunciationChallengeController');
-const SpeechChallenge = require('./administrative-sdk/models/speechChallenge');
-const SpeechChallengeController = require('./administrative-sdk/controllers/speechChallengeController');
-const SpeechRecording = require('./administrative-sdk/models/speechRecording');
-const SpeechRecordingController = require('./administrative-sdk/controllers/speechRecordingController');
+const PronunciationAnalysis = require('./administrative-sdk/models/pronunciation-analysis');
+const PronunciationAnalysisController = require('./administrative-sdk/controllers/pronunciation-analysis-controller');
+const PronunciationChallenge = require('./administrative-sdk/models/pronunciation-challenge');
+const PronunciationChallengeController = require('./administrative-sdk/controllers/pronunciation-challenge-controller');
+const SpeechChallenge = require('./administrative-sdk/models/speech-challenge');
+const SpeechChallengeController = require('./administrative-sdk/controllers/speech-challenge-controller');
+const SpeechRecording = require('./administrative-sdk/models/speech-recording');
+const SpeechRecordingController = require('./administrative-sdk/controllers/speech-recording-controller');
 const Student = require('./administrative-sdk/models/student');
-const StudentController = require('./administrative-sdk/controllers/studentController');
+const StudentController = require('./administrative-sdk/controllers/student-controller');
 const Tenant = require('./administrative-sdk/models/tenant');
-const TenantController = require('./administrative-sdk/controllers/tenantController');
+const TenantController = require('./administrative-sdk/controllers/tenant-controller');
 const Word = require('./administrative-sdk/models/word');
 const WordChunk = require('./administrative-sdk/models/word-chunk');
 const audioSdk = require('./audio-sdk');
