+++ test/choice-challengeSpec.js
@@ -1,7 +1,7 @@
 require('jasmine-ajax');
-const ChoiceChallenge = require('../administrative-sdk/models/choiceChallenge');
-const ChoiceChallengeController = require('../administrative-sdk/controllers/choiceChallengeController');
-const Connection = require('../administrative-sdk/controllers/connectionController');
+const ChoiceChallenge = require('../administrative-sdk/models/choice-challenge');
+const ChoiceChallengeController = require('../administrative-sdk/controllers/choice-challenge-controller');
+const Connection = require('../administrative-sdk/controllers/connection-controller');
 
 describe('ChoiceChallenge object test', () => {
   it('should require all required fields in constructor', () => {
